Tnayin2025

+++ src/room/Task28.java
@@ -2,7 +2,7 @@ package room;
 
 public class Task28 {
     public static void main(String[] args) {
-        int x = 2, y = 4, z = 6;
+        double x = 2, y = 4, z = 6;
         if (y / x == z / y) {
             System.out.println(true);
         } else {
